ktps - boj1016

+++ src/main/boj_ps/gold1/boj1016.kt
@@ -1,5 +1,5 @@
 package gold1
 
 fun main() = with(System.`in`.bufferedReader()) {
-    var (min, max) = readLine().split(" ").map { it.toInt() }
+    var (min, max) = readLine().split(" ").map { it.toLong() }
 }
